refactor(routing): replaced reactRouter.ts with reactRouter.tsx
