+++ src/client/src/routing/reactRouter.tsx
@@ -1,0 +1,35 @@
+import { createBrowserRouter } from "react-router";
+
+import RootLayout from "./RootLayout";
+import Home from "../pages/Home";
+import NotFound from "../pages/NotFound";
+import Search from "../pages/Search";
+import ErrorBoundary from "../pages/ErrorBoundary";
+import DashboardLayout from "../pages/dashboard/DashboardLayout";
+import Overview from "../pages/dashboard/sections/Overview";
+
+const router = createBrowserRouter([
+   {
+      id: "root",
+      path: "/",
+      Component: RootLayout,
+      errorElement: <ErrorBoundary />,
+      children: [
+         {
+            id: "home",
+            index: true,
+            Component: Home,
+         },
+         {
+            id: "dashboard",
+            path: "dashboard",
+            Component: DashboardLayout,
+            children: [{ index: true, Component: Overview }],
+         },
+         { id: "search", path: "search", Component: Search },
+         { id: "not-found", path: "*", Component: NotFound },
+      ],
+   },
+]);
+
+export default router;
